Output and input ammend

+++ src/graph/graph.cpp
@@ -833,7 +833,7 @@ vector<list<int>> Graph::hillClimbing(const int iteration_number, vector<list<in
 }
 
 
-vector<list<int>> Graph::simulatedAnnealing(const int iteration_number, const int coldness_ratio, vector<list<int>> (Graph::*mutation_func)(const vector<list<int>>&), int (Graph::*evaluation_func)(const vector<list<int>> &), bool log) {
+vector<list<int>> Graph::simulatedAnnealing(const int iteration_number, const float coldness_ratio, vector<list<int>> (Graph::*mutation_func)(const vector<list<int>>&), int (Graph::*evaluation_func)(const vector<list<int>> &), bool log) {
     vector<list<int>> best_solution = this->generate_closest_solution();
     int best_score = (this->*evaluation_func)(best_solution);
